user_service/requirements.txt
Django==5.1.8
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
drf-yasg==1.21.10
python-decouple==3.8
django-prometheus
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-exporter-otlp
